+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2025-06-24 21:48:07 - Bump version to $(node -p 'require(\'./package.json\').version')
+
+### Changed Files:
+- `custom-fields-block.php`
+
 ## 2025-06-24 21:43:23 - 1.2.4
 
 ### Changed Files: